Feat: day 50 binary search
- 백준 2295 세 수의 합: https://www.acmicpc.net/problem/2295
- 팩키지 구조 변경
