+++ src/fastcampus/part1/chapter07/binary_search/bj14425/Main2.java
@@ -1,4 +1,4 @@
-package fastcampus.part1.chapter7.binary_search.bj14425;
+package fastcampus.part1.chapter07.binary_search.bj14425;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -9,9 +9,9 @@ import java.util.StringTokenizer;
 /**
  * 백준 14425 문자열 집합
  * https://www.acmicpc.net/problem/14425
- * 1. Binary search를 이용한 풀이
+ * 3. Arrays의 binarySearch() 메서드를 이용한 풀이
  */
-public class Main {
+public class Main2 {
     static String[] setS;
 
     public static void main(String[] args) throws IOException {
@@ -23,33 +23,16 @@ public class Main {
         for (int i = 0; i < N; i++) {
             setS[i] = br.readLine();
         }
+        // binary search를 하기 전에 정렬을 먼저해야 하는 것을 잊지 말자.
         Arrays.sort(setS);
 
-        // setS에서 검사해야 하는 문자열이 있는지 찾는다.
         int answer = 0;
         for (int i = 0; i < M; i++) {
             String q = br.readLine();
-            if (isExistStr(q)) {
+            if (Arrays.binarySearch(setS, q) >= 0)  {
                 answer++;
             }
         }
         System.out.println(answer);
-    }
-
-    private static boolean isExistStr(String q) {
-        int left = 0;
-        int right = setS.length - 1;
-        while (left <= right) {
-            int m = (left + right) / 2;
-            int compareResult = setS[m].compareTo(q);
-            if (compareResult < 0) {
-                left = m + 1;
-            } else if (compareResult > 0) {
-                right = m - 1;
-            } else {
-                return true;
-            }
-        }
-        return false;
     }
 }
